requirements.txt
argh==0.30.4
beautifultable==1.1.0
blinker==1.6.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.6
consultor==0.2.0
dnspython==2.4.2
Flask==3.0.0
Flask-Resize==2.0.4
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pilkit==3.0
Pillow==10.1.0
pymongo==4.6.1
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.0
